requirements.txt
# --- DATA & ML PURE ---
pandas
yfinance
scikit-learn
joblib
psycopg2-binary
python-dotenv
requests

# --- BACKEND SAAS ---
fastapi
uvicorn
sqlalchemy

# --- AGENT MACRO & NLP (Gratuit) ---
langchain
langchain-google-genai  # <--- Le connecteur pour l'API gratuite Gemini
chromadb